requirements.txt
gunicorn==19.6.0
Flask==0.11.1
Flask-Script==2.0.5
Flask-Migrate==2.0.0
Flask-SQLAlchemy==2.1
Flask_RESTful==0.3.5
Flask-Responses==0.2
flask-cors==3.0.0
flask_marshmallow==0.7.0
marshmallow_sqlalchemy==0.13.0
psycopg2==2.6.2
equests==2.13.0
